requirements.txt
# Core LiveKit Agents SDK
livekit-agents~=1.4

# === STT/LLM/TTS Provider Plugins ===
# Google: Gemini LLM + Cloud STT + Cloud TTS + Gemini TTS
livekit-plugins-google~=1.4

# ElevenLabs: STT + TTS (install even if not using now — swappable)
livekit-plugins-elevenlabs~=1.4

# OpenAI: LLM + TTS + STT (also used for Groq via base_url)
livekit-plugins-openai~=1.4

# Anthropic: Claude LLM
livekit-plugins-anthropic~=1.4

# Cartesia: Ultra-low latency TTS
livekit-plugins-cartesia~=1.4

# Deepgram: STT (best for Bangla streaming)
livekit-plugins-deepgram~=1.4

# === VAD & Turn Detection ===
# Silero: Voice Activity Detection (local, free)
livekit-plugins-silero~=1.4

# Turn Detector: Semantic end-of-turn detection (local, free)
livekit-plugins-turn-detector~=1.4

# === Utilities ===
python-dotenv>=1.0
aiohttp>=3.9
